+++ CHANGES.rst
@@ -5,6 +5,9 @@ Version 3.1.5
 
 Unreleased
 
+-   Calling sync ``render`` for an async template uses ``asyncio.run``.
+    :pr:`1952`
+
 
 Version 3.1.4
 -------------
@@ -138,9 +141,8 @@ Released 2021-05-18
     extensions shows more relevant context. :issue:`1429`
 -   Fixed calling deprecated ``jinja2.Markup`` without an argument.
     Use ``markupsafe.Markup`` instead. :issue:`1438`
--   Calling sync ``render`` for an async template uses ``asyncio.run``
-    on Python >= 3.7. This fixes a deprecation that Python 3.10
-    introduces. :issue:`1443`
+-   Calling sync ``render`` for an async template uses ``asyncio.new_event_loop``
+    This fixes a deprecation that Python 3.10 introduces. :issue:`1443`
 
 
 Version 3.0.0