+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 -   Calling sync ``render`` for an async template uses ``asyncio.run``.
     :pr:`1952`
 -   Avoid unclosed ``auto_aiter`` warnings. :pr:`1960`
+-   Return an ``aclose``-able ``AsyncGenerator`` from
+    ``Template.generate_async``. :pr:`1960`
 
 
 Version 3.1.4